Incorporating review comments and minor clean-up
